backend/requirements.txt
# PQC Shield Backend
# Python 3.12+

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Auth & API
bcrypt>=4.1.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Redis, HTTP, optional AI
redis>=5.0.0
httpx>=0.26.0
anthropic>=0.18.0
groq>=0.4.0

# PQC (optional for tests: KEM/signature round-trips run if installed)
liboqs-python>=0.10.0

# PDF compliance report
reportlab>=4.0.0

# Billing
stripe>=8.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
